requirements.txt
altair==4.0.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrbox==0.1.5
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
Faker==22.6.0
fastjsonschema==2.18.0
favicon==0.7.0
filelock==3.13.1
firebase-admin==6.4.0
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googletrans==3.1.0a0
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
htbuilder==0.6.2
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
image==1.5.33
imageio==2.33.1
importlib-metadata==6.11.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lazy_loader==0.3
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.26.0
opencv-python-headless
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pdfmerge==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.7
Pympler==1.0.1
pyparsing==3.1.1
pypdf==3.9.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
semver==3.0.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
st-annotated-text==4.0.1
stack-data==0.6.2
streamlit==1.31.0
streamlit-authenticator==0.1.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
tifffile==2024.1.30
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.37.2
txtai==6.3.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
validators==0.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
widgetsnbextension==4.0.9
zipp==3.17.0
